added interface for each service class and changed tests
