+++ battle-service/src/main/java/by/pashkevich/mikhail/service/FieldVerifyService.java
@@ -1,0 +1,10 @@
+package by.pashkevich.mikhail.service;
+
+import by.pashkevich.mikhail.model.entity.Field;
+import by.pashkevich.mikhail.model.entity.enums.Value;
+
+public interface FieldVerifyService {
+    boolean isWin(Field field, Value value);
+
+    boolean isCorrect(Field field);
+}
