+++ battle-service/src/test/java/by/pashkevich/mikhail/service/impl/FieldServiceImplTest.java
@@ -1,11 +1,10 @@
-package by.pashkevich.mikhail.service;
+package by.pashkevich.mikhail.service.impl;
 
 import by.pashkevich.mikhail.model.entity.Field;
 import by.pashkevich.mikhail.model.entity.enums.BattleStatus;
 import by.pashkevich.mikhail.model.entity.enums.Value;
 import by.pashkevich.mikhail.repository.FieldRepository;
 import by.pashkevich.mikhail.service.data.MoveArgumentsProvider;
-import by.pashkevich.mikhail.service.util.field.FieldVerifyService;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.junit.jupiter.params.ParameterizedTest;
@@ -19,15 +18,15 @@ import org.mockito.junit.jupiter.MockitoExtension;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
 @ExtendWith(MockitoExtension.class)
-public class FieldServiceTest {
+public class FieldServiceImplTest {
     @Mock
     private FieldRepository fieldRepository;
 
     @Mock(answer = Answers.CALLS_REAL_METHODS)
-    private FieldVerifyService fieldVerifyService;
+    private FieldVerifyServiceImpl fieldVerifyService;
 
     @InjectMocks
-    private FieldService fieldService;
+    private FieldServiceImpl fieldService;
 
 
     @ParameterizedTest
