+++ battle-service/src/test/java/by/pashkevich/mikhail/service/impl/FieldVerifyServiceImplTest.java
@@ -1,22 +1,23 @@
-package by.pashkevich.mikhail.service.util.field;
+package by.pashkevich.mikhail.service.impl;
 
 import by.pashkevich.mikhail.model.entity.Field;
 import by.pashkevich.mikhail.model.entity.enums.Value;
-import by.pashkevich.mikhail.service.util.field.data.IsCorrectArgumentsProvider;
-import by.pashkevich.mikhail.service.util.field.data.IsWinArgumentsProvider;
+import by.pashkevich.mikhail.service.FieldVerifyService;
+import by.pashkevich.mikhail.service.data.IsCorrectArgumentsProvider;
+import by.pashkevich.mikhail.service.data.IsWinArgumentsProvider;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ArgumentsSource;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
-public class FieldVerifyServiceTest {
+public class FieldVerifyServiceImplTest {
     private static FieldVerifyService fieldVerifyService;
 
 
     @BeforeAll
     public static void beforeAll() {
-        fieldVerifyService = new FieldVerifyService();
+        fieldVerifyService = new FieldVerifyServiceImpl();
     }
 
     @ParameterizedTest
